feat: Implement ASCII and RTU APIs for Modbus operations
- Added ASCII-specific pure function API in `ascii.ts` for Modbus operations.
- Introduced RTU-specific pure function API in `rtu.ts` for Modbus operations.
- Removed the unified pure function API from `pure-functions.ts`.
- Updated frame parsing logic in `frameParser.ts` to support ASCII frames.
- Created comprehensive tests for ASCII API in `ascii-api.test.ts`.
- Adjusted integration tests to utilize the new RTU and ASCII APIs.
- Updated existing tests to reflect changes in API structure and functionality.
